Creating a base class for hub and iis controllers.
